Change `reproc_exit_status` return type to `int`.
This allows using `-1` as the value for when the process has not yet
exited. It also integrates better with returning child process exit
codes from the parent process as the return type of `main` is also
`int`.

+++ reproc/examples/drain.c
@@ -61,5 +61,5 @@ cleanup:
     return fail(error);
   }
 
-  return (int) reproc_exit_status(git_help);
+  return reproc_exit_status(git_help);
 }
